==> Sao Paulo/2_Part_notebook_stations_schedule.ipynb <==
import os

# Change directory to your target path
#os.chdir('/Users/carolinarutilidelima/Documents/OD_Sao_Paulo/')

os.chdir('/home/carolima/Documents/GitHub/OD_Sao_Paulo')


import pandas as pd

# Load the CSV file into a DataFrame
file_path = './output_data/top_30_station_pairs.csv'
df = pd.read_csv(file_path)

# Display the first few rows of the DataFrame to confirm it was loaded correctly
df.head()

from geopy.distance import geodesic

# Calculate distance between source and destination
def calculate_distance(row):
    source_coords = (row["source_lat"], row["source_lon"])
    dest_coords = (row["dest_lat"], row["dest_lon"])
    return geodesic(source_coords, dest_coords).kilometers

# Add a new column for distance
df["distance_km"] = df.apply(calculate_distance, axis=1)

# Filter out rows where the distance is less than 30km
filtered_df = df[df["distance_km"] >= 30]

# Display the filtered dataframe
filtered_df

import googlemaps
import populartimes


API_KEY = 'APIKEY'
gmaps = googlemaps.Client(key=API_KEY)


def get_popular_times_for_location(lat, lon):
    # Use the Places API to find nearby places with a larger radius
    places_result = gmaps.places_nearby(location=(lat, lon), radius=2000, type='establishment')
    
    # Iterate through the results to find a place with Popular Times data
    for place in places_result['results']:
        place_id = place['place_id']
        print(f"Checking Place ID: {place_id}")
        
        # Use populartimes library to get the popular times
        try:
            popular_times_data = populartimes.get_id(API_KEY, place_id)
            # Extract popular times information if available
            if 'populartimes' in popular_times_data:
                # Aggregate popularity data across all days
                hourly_popularity = [0] * 24
                for day_data in popular_times_data['populartimes']:
                    for hour, popularity in enumerate(day_data['data']):
                        hourly_popularity[hour] += popularity
                
                # Sort the hours based on aggregated popularity in descending order
                top_5_hours = sorted(range(24), key=lambda h: hourly_popularity[h], reverse=True)[:5]
                return top_5_hours
        except Exception as e:
            print(f"An error occurred while retrieving popular times data for place ID {place_id}: {e}")
    
    return None  # If no popular times data found



# Assuming df is the DataFrame loaded with the top 5 pairs data
output_data = []

for _, row in df.iterrows():  # limit to top 5 rows as per your requirement
    source_name = row['closest_source_station_ID']
    source_lat = row['source_lat']
    source_lon = row['source_lon']

    dest_name = row['closest_dest_station_ID']
    dest_lat = row['dest_lat']
    dest_lon = row['dest_lon']
    
    # Get the top 5 popular hours for this source
    top_5_source_hours = get_popular_times_for_location(source_lat, source_lon)

    # Get the top 5 popular hours for this destination
    top_5_dest_hours = get_popular_times_for_location(dest_lat, dest_lon)
    
    record = {
        'source': source_name,
        'source_lat': source_lat,
        'source_lon': source_lon,
        'destination': dest_name,
        'dest_lat': dest_lat,
        'dest_lon': dest_lon
    }

    if top_5_source_hours:
        for i, hour in enumerate(top_5_source_hours):
            record[f'source_hour{i+1}'] = hour
    
    if top_5_dest_hours:
        for i, hour in enumerate(top_5_dest_hours):
            record[f'dest_hour{i+1}'] = hour
    
    output_data.append(record)

# Create a DataFrame from the collected data
output_df_time = pd.DataFrame(output_data)

output_df_time

# Optionally, save to a CSV file
output_df_time.to_csv('output_data/output_df_time.csv', index=False)
