==> Sao Paulo/3_Part_notebook_stations_schedule.ipynb <==
import os

# Change directory to your target path
#os.chdir('/Users/carolinarutilidelima/Documents/OD_Sao_Paulo/')

os.chdir('/home/carolima/Documents/GitHub/OD_Sao_Paulo')


import pandas as pd

# Load the CSV file into a DataFrame
file_path = 'output_data/output_df_time.csv'
df = pd.read_csv(file_path)

# Display the first few rows of the DataFrame to confirm it was loaded correctly
df.head()


# Function to compute matching hours and select up to 5
def compute_matching_hours(row):
    source_hours = [
        row['source_hour1'], row['source_hour2'], row['source_hour3'],
        row['source_hour4'], row['source_hour5']
    ]
    dest_hours = [
        row['dest_hour1'], row['dest_hour2'], row['dest_hour3'],
        row['dest_hour4'], row['dest_hour5']
    ]
    
    # Remove NaNs
    source_hours = [hour for hour in source_hours if not pd.isna(hour)]
    dest_hours = [hour for hour in dest_hours if not pd.isna(hour)]
    
    # Find matching hours
    matching_hours = list(set(source_hours) & set(dest_hours))
    
    # If there are fewer than 5 matching hours, add popular non-matching hours
    if len(matching_hours) < 5:
        combined_hours = list(set(source_hours + dest_hours))
        combined_hours.sort()  # Sort for consistent results
        selected_hours = list(set(matching_hours))
        for hour in combined_hours:
            if hour not in selected_hours:
                selected_hours.append(hour)
            if len(selected_hours) == 5:
                break
    else:
        # Select only up to 5 matching hours
        selected_hours = matching_hours[:5]
    
    return selected_hours

# Apply the function to each row
df['selected_hours'] = df.apply(compute_matching_hours, axis=1)




df

# Generate the table with taking off station, landing station, and taking off time

output_rows = []



for _, row in df.iterrows():

    source_id = row['source']

    destination_id = row['destination']

    selected_hours = row['selected_hours']

    

    # Add a row for each selected hour

    for hour in selected_hours:

        output_rows.append({

            'taking_off_time': f"{int(hour):02}:00",  # Format hour to HH:00

            'taking_off_station_ID': int(source_id),

            'landing_station_ID': int(destination_id)

        })



# Create a new DataFrame for the output

output_df = pd.DataFrame(output_rows)


# Add index as a new column
output_df['index_number'] = output_df.index

# Reorder columns to make 'index_number' the first column
output_df = output_df[['index_number'] + [col for col in output_df.columns if col != 'index_number']]

# Save the updated DataFrame to a file
output_df.to_csv('output.csv', index=False)
